test: split test
